detail task

+++ app/Http/Controllers/TasksController.php
@@ -6,6 +6,12 @@ use Illuminate\Http\Request;
 use App\Task;
 class TasksController extends Controller
 {
+    public function detailTask(Request $request){
+        if($request->ajax()){
+            $task = Task::findOrFail($request->task_id);
+        }
+        return response()->json($task);
+    }
     public function createTask(Request $request){
         if($request->ajax()){
 //            return $request->all();
@@ -13,16 +19,17 @@ class TasksController extends Controller
             $task->title = $request->name;
             $task->board_id = $request->board_id;
             $task->save();
-            return response()->json($task);
         }
+        return response()->json($task);
     }
     public function updateTask(Request $request){
 
     }
     public function deleteTask(Request $request){
-//        return $request->all();
-        $task = Task::findOrFail($request->task_id);
-        $task->delete();
+        if($request->ajax()) {
+            $task = Task::findOrFail($request->task_id);
+            $task->delete();
+        }
         return response()->json($task);
     }
 }
